Add files via upload
The Death Recap Screen is a window displayed following the death of the player's character. It provides additional information on the cause of death, including the last several sources of damage dealt to the player.

It can be accessed by the recap button next to the release button while dead. It can also be accessed from clicking [You Died] from the chat log.

Extracted from https://github.com/bkader/KPack